+++ Test/tests/current_handler_test.c
@@ -1,6 +1,6 @@
 #include <string.h>
 
-#include "unity_fixture.h"      /* UNITY */
+#include <unity_fixture.h>      /* UNITY */
 #include "current_handler.h"    /* CUT */
 #include "logger.h"             /* CUT */
 #include "stm32l412xx-bsp.h"    /* CUT */
@@ -182,7 +182,7 @@ TEST ( Current_Handler, NormalToHighPrintout )
     char* expected = "High Current 35 dA\n";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -197,7 +197,7 @@ TEST ( Current_Handler, NormalToErrorPrintout )
     char* expected = "Error Current 40 dA\n";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -212,7 +212,7 @@ TEST ( Current_Handler, NormalToNormalNoPrintout )
     char* expected = "";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -230,7 +230,7 @@ TEST ( Current_Handler, HighToNormalPrintout )
     char* expected = "Normal Current 30 dA\n";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -248,7 +248,7 @@ TEST ( Current_Handler, HighToErrorPrintout )
     char* expected = "Error Current 40 dA\n";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -268,7 +268,7 @@ TEST ( Current_Handler, HighToHighNoPrintout )
     char* expected = "";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -286,7 +286,7 @@ TEST ( Current_Handler, ErrorToNormalPrintout )
     char* expected = "Normal Current 30 dA\n";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -304,7 +304,7 @@ TEST ( Current_Handler, ErrorToHighPrintout )
     char* expected = "High Current 35 dA\n";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -322,7 +322,7 @@ TEST ( Current_Handler, ErrorToErrorNoPrintout )
     char* expected = "";
     char* string = ( char* ) calloc ( 100, sizeof ( char ) );
 
-    ReadLog ( address, string, strlen ( expected ) );
+    ReadLog ( address, string, ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     free ( string );
@@ -338,7 +338,7 @@ TEST ( Current_Handler, LogCurrent )
     LogCurrent();
 
     ReadLog ( STARTING_CURRENT_ADDRESS + numCurrentLogs * CURRENT_LOG_SIZE, string,
-              strlen ( expected ) );
+              ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     numCurrentLogs++;
@@ -361,7 +361,7 @@ TEST ( Current_Handler, LogCurrentAgain )
     LogCurrent();
 
     ReadLog ( STARTING_CURRENT_ADDRESS + numCurrentLogs * CURRENT_LOG_SIZE, string,
-              strlen ( expected ) );
+              ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
 
     numCurrentLogs++;
@@ -391,7 +391,7 @@ TEST ( Current_Handler, LogFiftyCurrents )
     {
         snprintf ( expected, sizeof ( expected ), "%hu", i );
         ReadLog ( STARTING_CURRENT_ADDRESS + numCurrentLogs * CURRENT_LOG_SIZE, string,
-                  strlen ( expected ) );
+                  ( const uint32_t ) strlen ( expected ) );
         TEST_ASSERT_EQUAL_STRING ( expected, string );
         numCurrentLogs++;
 
@@ -448,7 +448,8 @@ TEST ( Current_Handler, WriteLogOverflow )
     LogCurrent();
 
     // Make sure that last log was written at STARTING_CURRENT_ADDRESS
-    ReadLog ( STARTING_CURRENT_ADDRESS, string, strlen ( expected ) );
+    ReadLog ( STARTING_CURRENT_ADDRESS, string,
+              ( const uint32_t ) strlen ( expected ) );
     TEST_ASSERT_EQUAL_STRING ( expected, string );
     numCurrentLogs++;
